Improved error handling
