+++ src/mapray/HTTP.js
@@ -1,4 +1,3 @@
-
 class HTTP {
 
     static get( url, option={} )
@@ -24,9 +23,12 @@ class HTTP {
             fetch( url, option )
             .then( response => {
                     if ( !response.ok ) {
-                        throw new FetchError( "Failed to Fetch: " + response.statusText, response );
+                        throw new FetchError( "Failed to fetch: " + response.statusText, response );
                     }
                     return response;
+            } )
+            .catch( error => {
+                    throw new FetchError( "Failed to fetch: ", null, error );
             } )
         );
     }
@@ -67,4 +69,21 @@ HTTP.CREDENTIAL_MODE = {
 
 };
 
+class FetchError extends Error {
+    constructor( message, response, cause )
+    {
+        super( message );
+        if ( Error.captureStackTrace ) {
+            Error.captureStackTrace( this, FetchError );
+        }
+        this.name = "FetchError";
+        this.response = response;
+        this.cause = cause;
+        if ( cause ) {
+            this.stack += "\nCaused-By: " + cause.stack;
+        }
+    }
+}
+
+export { FetchError };
 export default HTTP;
